eliminate duplicated javascript by using 'partials'
