+++ lib/run_loop.rb
@@ -25,6 +25,7 @@ require 'run_loop/host_cache'
 require 'run_loop/patches/awesome_print'
 require 'run_loop/core_simulator'
 require 'run_loop/simctl/plists'
+require 'run_loop/template'
 
 module RunLoop
