+++ lib/run_loop/core.rb
@@ -15,15 +15,15 @@ module RunLoop
     START_DELIMITER = "OUTPUT_JSON:\n"
     END_DELIMITER="\nEND_OUTPUT"
 
-    SCRIPTS_PATH = File.expand_path(File.join(File.dirname(__FILE__), '..', '..', 'scripts'))
     SCRIPTS = {
-        :dismiss => 'run_dismiss_location.js',
-        :run_loop_fast_uia => 'run_loop_fast_uia.js',
-        :run_loop_shared_element => 'run_loop_shared_element.js',
-        :run_loop_host => 'run_loop_host.js',
-        :run_loop_basic => 'run_loop_basic.js'
+      :dismiss => 'run_dismiss_location.js',
+      :run_loop_host => 'run_loop_host.js',
+      :run_loop_fast_uia => 'run_loop_fast_uia.js',
+      :run_loop_shared_element => 'run_loop_shared_element.js',
+      :run_loop_basic => 'run_loop_basic.js'
     }
 
+    SCRIPTS_PATH = File.expand_path(File.join(File.dirname(__FILE__), '..', '..', 'scripts'))
     READ_SCRIPT_PATH = File.join(SCRIPTS_PATH, 'read-cmd.sh')
     TIMEOUT_SCRIPT_PATH = File.join(SCRIPTS_PATH, 'timeout3')
 
@@ -54,7 +54,7 @@ module RunLoop
       if SCRIPTS[key].nil?
         return nil
       end
-      File.join(scripts_path, SCRIPTS[key])
+      SCRIPTS[key]
     end
 
     def self.detect_connected_device
@@ -244,7 +244,7 @@ module RunLoop
 
       script = File.join(results_dir, '_run_loop.js')
 
-      code = File.read(options[:script])
+      code = UIAScriptTemplate.new(SCRIPTS_PATH, options[:script]).result
       code = code.gsub(/\$PATH/, results_dir)
       code = code.gsub(/\$READ_SCRIPT_PATH/, READ_SCRIPT_PATH)
       code = code.gsub(/\$TIMEOUT_SCRIPT_PATH/, TIMEOUT_SCRIPT_PATH)
